Set gist URL and title via push options (#216)
